feat : 게시글 주인만 수정/삭제 가능하도록

+++ src/app/api/posts/[postId]/route.ts
@@ -1,4 +1,5 @@
 import { Post } from "@/types/Post";
+import { getAuthUesrOnServer } from "@/utils/getAuthUesrOnServer";
 import { createClient } from "@/utils/supabase/server";
 import { NextRequest, NextResponse } from "next/server";
 
@@ -21,7 +22,7 @@ export async function GET(_: NextRequest, params: parameter) {
   const supabase = createClient();
   const { data: post, error } = await supabase
     .from(TABLE_NAME)
-    .select("*, author: author_id(nickname, profile_image)")
+    .select("*, author: author_id(user_id, nickname, profile_image)")
     .eq(PK_COLUMN_NAME, id)
     .single();
 
@@ -33,32 +34,15 @@ export async function GET(_: NextRequest, params: parameter) {
 }
 
 /**
- * 게시글 Create
- * @returns post 객체
- */
-export async function POST(request: NextRequest) {
-  const { title, content, nickname, hashtags, image }: Post<false> = await request.json();
-
-  const supabase = createClient();
-  const { data: post } = await supabase
-    .from(TABLE_NAME)
-    .insert({
-      title,
-      content,
-      author_nickname: nickname,
-      hashtag: { tags: [...hashtags] },
-      image,
-    })
-    .select();
-
-  return NextResponse.json(post);
-}
-
-/**
  * 게시글 Update
  */
 export async function PUT(request: NextRequest, { params : {postId: id} }: parameter) {
-  const { title, content, hashtags, image } : Post<false> = await request.json();
+  const { title, content, authorId, hashtags, image } : Post<false> = await request.json();
+
+  const user = await getAuthUesrOnServer();
+  if (!user || authorId !== user.id) {
+    return NextResponse.json({error : {message : "수정할 수 없는 사용자입니다."}}, { status : 403 } );
+  }
 
   const supabase = createClient();
   const { data: post, error } =await supabase
@@ -70,20 +54,27 @@ export async function PUT(request: NextRequest, { params : {postId: id} }: param
       hashtag: { tags: hashtags },
     })
     .eq(PK_COLUMN_NAME, id)
-    .select();
+    .select()
+    .single();
   
-  return NextResponse.json(post);
+  return NextResponse.json({ data : post, error });
 }
 
 /**
  * 게시글 Delete
  */
 export async function DELETE(_: NextRequest, { params : {postId: id} }: parameter) {
+  const user = await getAuthUesrOnServer();
+  if (!user){
+    return NextResponse.json({error : {message : "로그인 후 이용해주세요"}}, { status : 403 } );
+  }
+  
   const supabase = createClient();
   const { error } = await supabase
     .from(TABLE_NAME)
     .delete()
-    .eq(PK_COLUMN_NAME, id);
+    .eq(PK_COLUMN_NAME, id)
+    .eq("author_id", user.id);
 
-  return NextResponse.json(error ? error : `게시글(${id}) 삭제 완료`);
+  return NextResponse.json({ error }, {status : (!error ? 200 : 403) });
 }
